tests/wpt/web-platform-tests/priority-hints/service-worker-fetch-request-priority.tentative.https.html

<!DOCTYPE html>
<html>
<title>Priority Hints: Service Worker Fetch Request Priority</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
<script>

let frame = null;
let worker = null;
const script = 'resources/service-worker-fetch-request-priority.js';
const scope = 'resources/blank.html';
const channel = new MessageChannel();

// Wait for a message from the service worker and removes the message handler.
function wait_for_message_from_worker() {
  return new Promise((resolve) => channel.port2.onmessage = (event) => resolve(event.data));
}

// Global setup: this must be the first promise_test.
promise_test(async (t) => {
  const registration =
      await service_worker_unregister_and_register(t, script, scope);
  worker = registration.installing;
  if (!worker)
      worker = registration.active;
  worker.postMessage({port:channel.port1}, [channel.port1]);
  await wait_for_state(t, worker, 'activated');
  frame = await with_iframe(scope);
}, 'global setup');

promise_test(async (t) => {
  frame.contentWindow.fetch('blank.html?priority', {priority: 'high'});
  const message = await wait_for_message_from_worker();
  assert_equals(message, 'high');
}, "'high' fetch priority is reflected in service worker fetch handler's "+
    " request object's priority as 'high'");

promise_test(async (t) => {
  frame.contentWindow.fetch('blank.html?priority', {priority: 'low'});
  const message = await wait_for_message_from_worker();
  assert_equals(message, 'low');
}, "'low' fetch priority is reflected in service worker fetch handler's "+
    " request object's priority as 'low'");

promise_test(async (t) => {
  frame.contentWindow.fetch('blank.html?priority', {priority: 'auto'});
  const message = await wait_for_message_from_worker();
  assert_equals(message, 'auto');
}, "'auto' fetch priority is reflected in service worker fetch handler's "+
    " request object's priority as 'auto'");

promise_test(async (t) => {
  frame.contentWindow.fetch('blank.html?priority');
  const message = await wait_for_message_from_worker();
  assert_equals(message, 'auto');
}, "default fetch priority is reflected in service worker fetch handler's "+
    " request object's priority as 'auto'");

promise_test(async (t) => {
  const doc = frame.contentDocument;
  const node = doc.createElement("img");
  node.src = "img-auto?priority";
  doc.body.appendChild(node);
  const message = await wait_for_message_from_worker();
  assert_equals(message, 'auto');
}, "img element with default fetchpriority is reflected in service worker" +
    " fetch handler's request object's priority as 'auto'");

promise_test(async (t) => {
  const doc = frame.contentDocument;
  const node = doc.createElement("img");
  node.fetchPriority = "high";
  node.src = "img-high?priority";
  doc.body.appendChild(node);
  const message = await wait_for_message_from_worker();
  assert_equals(message, 'high');
}, "img element with fetchpriority of high is reflected in service worker" +
    " fetch handler's request object's priority as 'high'");

promise_test(async (t) => {
  const doc = frame.contentDocument;
  const node = doc.createElement("script");
  node.src = "script-auto?priority";
  doc.body.appendChild(node);
  const message = await wait_for_message_from_worker();
  assert_equals(message, 'auto');
}, "script element with default fetchpriority is reflected in service worker" +
    " fetch handler's request object's priority as 'auto'");

promise_test(async (t) => {
  const doc = frame.contentDocument;
  const node = doc.createElement("script");
  node.fetchPriority = "high";
  node.src = "script-high?priority";
  doc.body.appendChild(node);
  const message = await wait_for_message_from_worker();
  assert_equals(message, 'high');
}, "script element with fetchpriority of high is reflected in service worker" +
    " fetch handler's request object's priority as 'high'");

promise_test(async (t) => {
  const doc = frame.contentDocument;
  const node = doc.createElement("link");
  node.rel = "stylesheet";
  node.type = "text/css";
  node.href = "link-auto?priority";
  doc.body.appendChild(node);
  const message = await wait_for_message_from_worker();
  assert_equals(message, 'auto');
}, "link element with default fetchpriority is reflected in service worker" +
    " fetch handler's request object's priority as 'auto'");

promise_test(async (t) => {
  const doc = frame.contentDocument;
  const node = doc.createElement("link");
  node.rel = "stylesheet";
  node.type = "text/css";
  node.fetchPriority = "high";
  node.href = "link-high?priority";
  doc.body.appendChild(node);
  const message = await wait_for_message_from_worker();
  assert_equals(message, 'high');
}, "link element with fetchpriority of high is reflected in service worker" +
    " fetch handler's request object's priority as 'high'");

promise_test(async (t) => {
  const frame = document.createElement('iframe');
  frame.src = scope + '?priority-default';
  frame.style.position = 'absolute';
  document.body.appendChild(frame);
  const message = await wait_for_message_from_worker();
  assert_equals(message, 'auto');
  frame.parentNode.removeChild(frame);
}, "iframe element with default fetchpriority is reflected in service worker" +
  " fetch handler's request object's priority as 'auto'");

promise_test(async (t) => {
  const frame = document.createElement('iframe');
  frame.fetchPriority = 'high';
  frame.src = scope + '?priority-high';
  frame.style.position = 'absolute';
  document.body.appendChild(frame);
  const message = await wait_for_message_from_worker();
  assert_equals(message, 'high');
  frame.parentNode.removeChild(frame);
}, "iframe element with fetchpriority of high is reflected in service worker" +
  " fetch handler's request object's priority as 'high'");

// Global cleanup: the final promise_test.
promise_test(async (t) => {
  if (frame)
    frame.remove();
  await service_worker_unregister(t, scope);
}, 'global cleanup');
</script>
</html>
